+++ api/controllers/api_commentController.js
@@ -26,6 +26,10 @@ exports.addComment = async (req, res) => {
   const collection_id = req.params.id;
   const { comment, user_id } = req.body;
 
+  if (user_id !== req.user.id) {
+    return res.status(403).json({ error: "Not authorised" });
+  }
+
   try {
     await Comment.create({
       collection_id: collection_id,
@@ -44,12 +48,21 @@ exports.deleteComment = async (req, res) => {
   const comment_id = req.params.comment_id;
 
   try {
+    const userVerify = await Comment.findOne({
+      where: {
+        user_id: req.user.id,
+        comment_id: comment_id,
+      },
+    });
+    if (!userVerify) {
+      return res.status(403).json({ error: "Not authorised" });
+    }
+
     await Comment.destroy({
       where: {
         comment_id: comment_id,
       },
     });
-
     return res.status(200).json({ success: "Comment deleted" });
   } catch (err) {
     console.log(err);
